requirements.txt
langchain
openai
gradio
pypdf
PyPDF2
faiss-cpu
tiktoken
pinecone-client